Fix a potential crash in ApiSectionResponse
There is a case where a notebook can have no sections in it. Under these
conditions the existing code would crash, as generating a new URL with a
null string throws.  So, checking to ensure that the property exists
before generating the ULR fixes the problem.

Also, there was some incorrect logic around the "links" node. The schema
does not allow for the pagesUrl property, so the code as written will
never work.  Cleaning out that logic fixes things up.